N19/requirements.txt
fastapi
uvicorn
openai-whisper
# torch and torchaudio are required for whisper
torch
 torchaudio
